@@ -123,7 +123,12 @@ declare module 'react-native-slider' {
     /**
      * Used to configure the animation parameters.  These are the same parameters in the Animated library.
      */
-    animationConfig?: SpringAnimationConfig | TimingAnimationConfig
+    animationConfig?: SpringAnimationConfig | TimingAnimationConfig,
+
+    /**
+     * Used to replace Image component.
+     */
+    iconComponent: any,
   }
 
   const Slider: ComponentClass<SliderProps>
